+++ CHANGELOG.md
@@ -14,15 +14,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   an extra path with modules.
 - `tt play`: support connection to a target instance by `application` name
   or `application:instance` name.
+- `tt coredump pack`: add options to customize coredump packing:
+  * `-e (--executable)`: specify Tarantool executable path.
+  * `-p (--pid)`: specify PID of the dumped process.
+  * `-t (--time)`: specify time of dump (seconds since the Epoch).
 
 ### Changed
 
 - `tt stop/kill/clean/logrotate`: no longer need:
    * Instances scripts for multi-instance applications.
    * Cluster config for tarantool3-based cluster applications.
-
 - `tt logrotate`: don't exit at non-running instance, just warn and proceed
   with the other instances, like `tt stop` and `tt kill` do.
+- `tt coredump pack`: by default tarantool executable path is obtained from
+  `PATH` instead of using the hardcoded path `/usr/bin/tarantool`.
 
 ### Fixed